test: split memos CLI run tests by operation into dedicated files
